requirements.txt
flet>=0.80.5
cryptography>=42.0.0
argon2-cffi>=23.1.0
mnemonic>=0.20
qrcode>=8.2
Pillow>=12.1.0
